adding sqlalchemy dep and refactoring run key.
