+++ space/aspi/problemspace.yaml
@@ -1,23 +1,3 @@
-run_key:
-  run_id:
-    required: True
-    default: null
-    verbose_name: Run Id
-    short_name: run_id
-    feature_type_str: int
-  timestamp_added:
-    required: False
-    default: 1996-09-11 00:00:00.000+02:00
-    verbose_name: Timestamp Added
-    short_name: ts_added
-    feature_type_str: datetime
-  added_from:
-    required: False
-    default: RUN
-    verbose_name: Added From
-    short_name: from
-    feature_type_str: str
-
 instance_key:
   # more thoughts here, potentially answering derived column potentially as well:
   # TODO: instances table, just a foreign key in every row in results table
