+++ tests/test_optispace.py
@@ -3,6 +3,11 @@ import pytest
 
 from pathlib import Path
 from typing import Any
+from optiface.core.featuredata import (
+    init_data_feature_run_id,
+    init_data_feature_timestamp_added,
+    init_data_feature_added_from,
+)
 from optiface.core.optidatetime import OptiDateTimeFactory
 from optiface.core.optispace import (
     yaml_to_feature_type,
@@ -34,39 +39,6 @@ _DEFAULT_PSPACE_NAME: str = "defaultproblem"
 _DEFAULT_PSPACE_PATH: Path = Path(_SPACE) / _DEFAULT_PSPACE_NAME / _PSPACE_YAML
 
 opti_dt = OptiDateTimeFactory()
-
-
-def init_data_feature_run_id() -> dict[str, Any]:
-    return {
-        "name": "run_id",
-        "required": True,
-        "default": None,
-        "verbose_name": "Run Id",
-        "short_name": "run_id",
-        "feature_type_str": "int",
-    }
-
-
-def init_data_feature_timestamp_added() -> dict[str, Any]:
-    return {
-        "name": "timestamp_added",
-        "required": False,
-        "default": opti_dt.optidefault(),
-        "verbose_name": "Timestamp Added",
-        "short_name": "ts_added",
-        "feature_type_str": "datetime",
-    }
-
-
-def init_data_feature_added_from() -> dict[str, Any]:
-    return {
-        "name": "added_from",
-        "required": False,
-        "default": "RUN",
-        "verbose_name": "Added From",
-        "short_name": "from",
-        "feature_type_str": "str",
-    }
 
 
 def init_data_feature_set_name() -> dict[str, Any]:
